test : mapperConfig 설정 추가
