+++ src/test/java/com/pser/hotel/domain/hotel/dao/HotelDaoTest.java
@@ -1,7 +1,12 @@
 package com.pser.hotel.domain.hotel.dao;
 
+import com.pser.hotel.domain.hotel.config.MapperConfig;
 import com.pser.hotel.domain.hotel.domain.Facility;
 import com.pser.hotel.domain.hotel.domain.Hotel;
+import com.pser.hotel.domain.hotel.dto.HotelMapper;
+import com.pser.hotel.domain.hotel.dto.HotelMapperImpl;
+import com.pser.hotel.domain.hotel.dto.TimesaleMapper;
+import com.pser.hotel.domain.hotel.dto.TimesaleMapperImpl;
 import com.pser.hotel.domain.hotel.util.Utils;
 import com.pser.hotel.domain.member.domain.User;
 import com.pser.hotel.global.config.QueryDslConfig;
@@ -10,15 +15,32 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@Import({QueryDslConfig.class})
+@Import({QueryDslConfig.class, MapperConfig.class})
 public class HotelDaoTest {
+
+    @TestConfiguration
+    static class TestConfig {
+        @Bean
+        public HotelMapper hotelMapper() {
+            return new HotelMapperImpl();
+        }
+
+        @Bean
+        public TimesaleMapper timesaleMapper() {
+            return new TimesaleMapperImpl();
+        }
+    }
+
     @Autowired
     HotelDao hotelDao;
     @Autowired
